Implement base scheduler abstract class, fix issue in the estimated finish time data type

+++ src/scheduler_heft.hpp
@@ -22,5 +22,5 @@ class HEFT_Scheduler : public EFT_Scheduler
 
     void print();
 
-    std::tuple<simgrid_exec_t *, int, unsigned long> next() override;
+    std::tuple<simgrid_exec_t *, int, double> next() override;
 };
